Regeneradas clases de EntityFramework

+++ cmh/terminal/CheekiBreeki.CMH.Terminal.DAL/PERSONAL.cs
@@ -25,7 +25,7 @@ namespace CheekiBreeki.CMH.Terminal.DAL
         public string APELLIDOS { get; set; }
         public Nullable<int> REMUNERACION { get; set; }
         public string HASHED_PASS { get; set; }
-        public Nullable<int> PORCENT_DESCUENTO { get; set; }
+        public Nullable<byte> PORCENT_DESCUENTO { get; set; }
     
         public virtual ICollection<FUNCIONARIO> FUNCIONARIO { get; set; }
         public virtual ICollection<PERS_MEDICO> PERS_MEDICO { get; set; }
